requirements.txt
streamlit==1.28.0
datasets==2.14.6
sentence-transformers==2.2.2
pinecone-client==3.0.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
langchain==0.1.0
langchain-community==0.0.10


